backend/requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
ultralytics>=8.0.0
flask>=2.0.0
flask-cors>=4.0.0 
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
torch>=2.0.0
facenet-pytorch>=2.5.0
pytz>=2023.3 